third_party/blink/web_tests/fast/html/imports/import-script-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Test that scripts in imported documents are executed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS givenDocument is document
PASS givenDocument.currentScript.ownerDocument is leaf.import
PASS labels is loadedLabels
PASS successfullyParsed is true

TEST COMPLETE
